Do not lint for PROMP|SOLUTION problems
I think doing this in js is a bad idea, and console shouldn't be used
in "production" (eslint doesn't like it). A bettter solution would be
writing some Ruby script, that checks all source files for any problems
with those special comments. The only drawback is that this *.rb script
would need to be invoked manually.